fix according to the pull request comment
