+++ config.yml
@@ -2,14 +2,14 @@ OSMfile: osmData/TX-To-TU.osm
 buildConnFile: buildingConnection.csv
 buildingConfigPath: config/tsukuba-tu-building-data.csv
 jobsFile: config/jobs.csv
-numberOfAgents: 5000
-threadNumber: 10
+numberOfAgents: 1000
+threadNumber: 4
 infectedAgent: 20
 vaccinationPercentage: 0.0
 verbose: false
 gridHeight: 10
 gridWidth: 10
-windowWidth: 45%
+windowWidth: 75%
 windowHeight: 75%
 reportDir: report
 reportInterval: 5
